real-estate-app/backend/requirements.txt
# Основні залежності для MVP
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0

# Для роботи з базами даних
psycopg2-binary==2.9.9

# Для HTTP запитів
requests==2.31.0
aiohttp==3.9.1

# Для парсингу HTML
beautifulsoup4==4.12.2
lxml==4.9.3

# Для геолокації
geopy==2.4.1

# Допоміжні інструменти
python-dotenv==1.0.0
click==8.1.7
